Added IO to loop and updated acts
This will be the final commit dealing with homemade loops, tasks, and everything else. There will be two branches, `legacy` and `master`, where legacy is the current project and master will be where all new work goes into. This new work will be mainly a reimplementation of curio's Kernel built on top of tkinter's loop. The legacy will need no curio, but the new master will.